+++ .craft.yml
@@ -112,6 +112,9 @@ targets:
     id: '@sentry/tanstackstart'
     includeNames: /^sentry-tanstackstart-\d.*\.tgz$/
   - name: npm
+    id: '@sentry/tanstackstart-react'
+    includeNames: /^sentry-tanstackstart-react-\d.*\.tgz$/
+  - name: npm
     id: '@sentry/gatsby'
     includeNames: /^sentry-gatsby-\d.*\.tgz$/
   - name: npm
