+++ dev-packages/e2e-tests/verdaccio-config/config.yaml
@@ -176,6 +176,12 @@ packages:
     unpublish: $all
     # proxy: npmjs # Don't proxy for E2E tests!
 
+  '@sentry/tanstackstart-react':
+    access: $all
+    publish: $all
+    unpublish: $all
+    # proxy: npmjs # Don't proxy for E2E tests!
+
   '@sentry/types':
     access: $all
     publish: $all
