+++ packages/tanstackstart-react/.eslintrc.js
@@ -1,0 +1,10 @@
+module.exports = {
+  env: {
+    browser: true,
+    node: true,
+  },
+  parserOptions: {
+    jsx: true,
+  },
+  extends: ['../../.eslintrc.js'],
+};
