+++ packages/tanstackstart-react/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@sentry/tanstackstart",
+  "name": "@sentry/tanstackstart-react",
   "version": "9.5.0",
-  "description": "Utilities for the Sentry TanStack Start SDKs",
+  "description": "Official Sentry SDK for TanStack Start React",
   "repository": "git://github.com/getsentry/sentry-javascript.git",
-  "homepage": "https://github.com/getsentry/sentry-javascript/tree/master/packages/tanstackstart",
+  "homepage": "https://github.com/getsentry/sentry-javascript/tree/master/packages/tanstackstart-react",
   "author": "Sentry",
   "license": "MIT",
   "engines": {
@@ -27,6 +27,16 @@
         "import": "./build/esm/index.server.js",
         "require": "./build/cjs/index.server.js"
       }
+    },
+    "./import": {
+      "import": {
+        "default": "./build/import-hook.mjs"
+      }
+    },
+    "./loader": {
+      "import": {
+        "default": "./build/loader-hook.mjs"
+      }
     }
   },
   "typesVersions": {
@@ -39,7 +49,15 @@
   "publishConfig": {
     "access": "public"
   },
-  "dependencies": {},
+  "dependencies": {
+    "@opentelemetry/api": "^1.9.0",
+    "@opentelemetry/semantic-conventions": "^1.30.0",
+    "@sentry-internal/browser-utils": "9.5.0",
+    "@sentry/core": "9.5.0",
+    "@sentry/node": "9.5.0",
+    "@sentry/opentelemetry": "9.5.0",
+    "@sentry/react": "9.5.0"
+  },
   "scripts": {
     "build": "run-p build:transpile build:types",
     "build:dev": "yarn build",
@@ -53,7 +71,7 @@
     "build:types:watch": "tsc -p tsconfig.types.json --watch",
     "build:tarball": "npm pack",
     "circularDepCheck": "madge --circular src/index.client.ts && madge --circular src/index.server.ts && madge --circular src/index.types.ts",
-    "clean": "rimraf build coverage sentry-tanstackstart-*.tgz",
+    "clean": "rimraf build coverage sentry-tanstackstart-react-*.tgz",
     "fix": "eslint . --format stylish --fix",
     "lint": "eslint . --format stylish",
     "test": "yarn test:unit",
