+++ packages/tanstackstart-react/test/temp.test.ts
@@ -1,0 +1,7 @@
+import { describe, it, expect } from 'vitest';
+
+describe('Basic test suite', () => {
+  it('should pass', () => {
+    expect(true).toBe(true);
+  });
+});
