+++ packages/tanstackstart-react/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src/**/*"],
+  "compilerOptions": {
+    "lib": ["es2018", "es2020.string"],
+    "module": "Node16"
+  }
+}
