+++ packages/tanstackstart-react/tsconfig.test.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.json",
+  "include": ["test/**/*", "vite.config.ts"],
+  "compilerOptions": {
+    "types": ["node"],
+    "lib": ["DOM", "ESNext"]
+  }
+}
