+++ packages/tanstackstart-react/vite.config.ts
@@ -1,0 +1,8 @@
+import baseConfig from '../../vite/vite.config';
+
+export default {
+  ...baseConfig,
+  test: {
+    ...baseConfig.test,
+  },
+};
