+++ packages/tanstackstart/src/client/index.ts
@@ -1,1 +1,1 @@
-export * from '@sentry/react';
+export {};
