+++ packages/tanstackstart/src/server/index.ts
@@ -1,42 +1,1 @@
-export * from '@sentry/node';
-
-/**
- * A passthrough error boundary for the server that doesn't depend on any react. Error boundaries don't catch SSR errors
- * so they should simply be a passthrough.
- */
-export const ErrorBoundary = (props: React.PropsWithChildren<unknown>): React.ReactNode => {
-  if (!props.children) {
-    return null;
-  }
-
-  if (typeof props.children === 'function') {
-    return (props.children as () => React.ReactNode)();
-  }
-
-  return props.children;
-};
-
-/**
- * A passthrough redux enhancer for the server that doesn't depend on anything from the `@sentry/react` package.
- */
-export function createReduxEnhancer() {
-  return (createStore: unknown) => createStore;
-}
-
-/**
- * A passthrough error boundary wrapper for the server that doesn't depend on any react. Error boundaries don't catch
- * SSR errors so they should simply be a passthrough.
- */
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function withErrorBoundary<P extends Record<string, any>>(
-  WrappedComponent: React.ComponentType<P>,
-): React.FC<P> {
-  return WrappedComponent as React.FC<P>;
-}
-
-/**
- * Just a passthrough since we're on the server and showing the report dialog on the server doesn't make any sense.
- */
-export function showReportDialog(): void {
-  return;
-}
+export {};
